+++ AllycsPointRuler/PointRulerProductModel.cs
@@ -6,8 +6,8 @@ namespace AllycsPointRuler
 {
     public class PointRulerProductModel
     {
-        public int TotalProducePoints { get; set; } = 0;
-        public int TotalProduceAmounts { get; set; } = 0;
+        public Int64 TotalProducePoints { get; set; } = 0;
+        public Int64 TotalProduceAmounts { get; set; } = 0;
         public List<CalMethodModel> CalMethods { get; set; } = new List<CalMethodModel>();
     }
 }
